Clean up remaining references to the Reference class in the core folder. It has been moved to children folder.
